+++ NEWS.md
@@ -32,3 +32,4 @@
 * Updated UK pop-age file for scenario 1 on server
 * Replaced `dplyr::summarize()` with `dplyr::reframe()` in `get_wcde_single()`
 * Add `version` argument in `get_wcde()`
+* Changed files on servers to `.rds` and code in `get_wcde()` to retrieve `.rds` data